requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cloudinary==1.38.0
crispy-bootstrap4==2024.10
dj-database-url==2.1.0
Django==4.2.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-crispy-forms==2.3
djangorestframework==3.14.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.4
multidict==6.0.4
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
pytz==2023.3.post1
pywatchman==1.4.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==8.1.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.0.7
whitenoise==6.6.0
yarl==1.9.4
